services/core_api/requirements.txt
fastapi[all]
uvicorn
sqlmodel
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
httpx
websockets
psycopg2-binary
python-dotenv

# AI & Voice Pipeline Requirements
openai-whisper
gTTS
langdetect
deep-translator
faiss-cpu
sentence-transformers
torch
torchaudio
